chore: model definitions moved to base file and payment_request model file deleted
