+++ tests/wallet/test_cli_nut26.py
@@ -3,8 +3,8 @@ import json
 import pytest
 from click.testing import CliRunner
 
+from cashu.core.base import NUT10Option, PaymentRequest
 from cashu.core.nuts.nut26 import serialize as nut26_serialize
-from cashu.core.nuts.payment_request import NUT10Option, PaymentRequest
 from cashu.core.settings import settings
 from cashu.wallet.cli.cli import cli
 from tests.helpers import is_fake
